+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-beta.29] - 2025-05-22
+
+### Added
+
+- Checkouts API endpoints:
+  - Update the checkout details (step 1)
+  - Update the shipping service (step 2)
+  - Process payment (step 3)
+  - Create a payment intent
+
 ## [1.0.0-beta.28] - 2025-05-20
 
 ### Added